新增： 1.支持HELM_CHART角色部署 2.当部署HELM_CONTROLLER组件时，注入kube_config 修改： 1.改进K8sUtil类，改进日志监听方式，监听创建过程容器日志 2.修改类名称为RoleK8sTask 3.优化部分代码

+++ cloudeon-server/src/main/java/org/dromara/cloudeon/processor/RoleK8sTask.java
@@ -21,13 +21,13 @@ import cn.hutool.core.io.IoUtil;
 import cn.hutool.extra.spring.SpringUtil;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.api.model.batch.v1.Job;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.ParameterNamespaceListVisitFromServerGetDeleteRecreateWaitApplicable;
-import io.fabric8.kubernetes.client.dsl.Resource;
-import io.fabric8.kubernetes.client.dsl.RollableScalableResource;
 import lombok.NoArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.dromara.cloudeon.config.CloudeonConfigProp;
+import org.dromara.cloudeon.crd.helmchart.HelmChart;
 import org.dromara.cloudeon.dao.*;
 import org.dromara.cloudeon.entity.*;
 import org.dromara.cloudeon.enums.ServiceRoleState;
@@ -40,6 +40,7 @@ import java.io.File;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -73,12 +74,16 @@ public abstract class RoleK8sTask extends BaseCloudeonTask implements ApplyOrDel
         // 查询框架服务角色名获取模板名
         String roleName = taskParam.getRoleName();
         StackServiceRoleEntity stackServiceRoleEntity = stackServiceRoleRepository.findByServiceIdAndName(taskParam.getStackServiceId(), roleName);
-        String roleFullName = stackServiceRoleEntity.getRoleFullName();
+        String roleFullName = K8sUtil.formatK8sNameStr(stackServiceRoleEntity.getRoleFullName());
 
         Integer serviceInstanceId = taskParam.getServiceInstanceId();
         String roleType = stackServiceRoleEntity.getType();
-        //当角色类型为EMPTY，不进行任何具体操作
-        if ("EMPTY".equalsIgnoreCase(roleType)) {
+
+        // 查询本服务实例拥有的指定角色节点数
+        int roleNodeCnt = serviceRoleInstanceRepository.countByServiceInstanceIdAndServiceRoleName(serviceInstanceId, roleName);
+
+        //当角色没有分配节点时，不执行部署
+        if ("EMPTY".equalsIgnoreCase(roleType) || roleNodeCnt == 0) {
             // 更新角色实例状态
             List<ServiceRoleInstanceEntity> roleInstanceEntities = serviceRoleInstanceRepository.findByServiceInstanceIdAndServiceRoleName(serviceInstanceId, stackServiceRoleEntity.getName());
             roleInstanceEntities.forEach(r -> {
@@ -92,7 +97,7 @@ public abstract class RoleK8sTask extends BaseCloudeonTask implements ApplyOrDel
         StackServiceEntity stackServiceEntity = stackServiceRepository.findById(serviceInstanceEntity.getStackServiceId()).get();
 
         String stackCode = stackServiceEntity.getStackCode();
-        String stackServiceName = stackServiceEntity.getName().toLowerCase();
+        String stackServiceName = stackServiceEntity.getName().toLowerCase().replace('_', '-');
 
         // 查询服务实例所有配置项
         List<ServiceInstanceConfigEntity> allConfigEntityList = configRepository.findByServiceInstanceId(serviceInstanceId);
@@ -112,21 +117,32 @@ public abstract class RoleK8sTask extends BaseCloudeonTask implements ApplyOrDel
         String k8sTemplateDir = cloudeonConfigProp.getStackLoadPath() + File.separator + stackCode + File.separator + stackServiceName + File.separator + Constant.K8S_DIR;
         log.info("加载服务实例角色k8s资源模板目录：" + k8sTemplateDir);
 
-        // 查询本服务实例拥有的指定角色节点数
-        int roleNodeCnt = serviceRoleInstanceRepository.countByServiceInstanceIdAndServiceRoleName(serviceInstanceId, roleName);
-
-
         // 构建数据模型
         Map<String, Object> dataModel = new HashMap<>();
-        String roleServiceFullName = roleFullName + "-" + serviceInstanceEntity.getServiceName().toLowerCase();
+        String roleServiceFullName = roleFullName + "-" + K8sUtil.formatK8sNameStr(serviceInstanceEntity.getServiceName());
         dataModel.put("roleFullName", roleFullName);
         dataModel.put("roleServiceFullName", roleServiceFullName);
-        dataModel.put("serviceFullName", serviceInstanceEntity.getServiceName().toLowerCase());
+        dataModel.put("serviceFullName", K8sUtil.formatK8sNameStr(serviceInstanceEntity.getServiceName()));
         dataModel.put("service", serviceInstanceEntity);
         dataModel.put("roleNodeCnt", roleNodeCnt);
         dataModel.put("runAs", stackServiceEntity.getRunAs());
         dataModel.put("conf", allConfigEntityList.stream().collect(Collectors.toMap(ServiceInstanceConfigEntity::getName, ServiceInstanceConfigEntity::getValue)));
         dataModel.put("namespace", namespace);
+        // 获取该服务支持的自定义配置文件名
+        String customConfigFiles = stackServiceEntity.getCustomConfigFiles();
+        Map<String, Map<String, String>> confFiles = new HashMap<>();
+
+        if (StringUtils.isNoneBlank(customConfigFiles)) {
+            for (String confFileName : customConfigFiles.split(",")) {
+                List<ServiceInstanceConfigEntity> groupConfEntities = configRepository.findByServiceInstanceIdAndConfFile(serviceInstanceId, confFileName);
+                HashMap<String, String> map = new HashMap<>();
+                for (ServiceInstanceConfigEntity groupConf : groupConfEntities) {
+                    map.put(groupConf.getName(), groupConf.getValue());
+                }
+                confFiles.put(confFileName, map);
+            }
+        }
+        dataModel.put("confFiles", confFiles);
 
         // 根据角色名称获取最合适的k8s模板文件
         String k8sTemplateFileName = "";
@@ -136,44 +152,63 @@ public abstract class RoleK8sTask extends BaseCloudeonTask implements ApplyOrDel
             }
         }
         String modelYamlStr = FreemarkerUtil.templateEval(FileUtil.readUtf8String(k8sTemplateDir + File.separator + k8sTemplateFileName), dataModel);
-
+        // 当生成内容为空白时，则不需执行
+        if (StringUtils.isBlank(modelYamlStr)) {
+            // 更新角色实例状态
+            List<ServiceRoleInstanceEntity> roleInstanceEntities = serviceRoleInstanceRepository.findByServiceInstanceIdAndServiceRoleName(serviceInstanceId, stackServiceRoleEntity.getName());
+            roleInstanceEntities.forEach(r -> {
+                r.setServiceRoleState(getTargetState());
+                serviceRoleInstanceRepository.save(r);
+            });
+            return;
+        }
         // 调用k8s命令启动资源
         KubernetesClient client = kubeService.getKubeClient(serviceInstanceEntity.getClusterId());
-        String resourceName = "";
 
         try {
             ParameterNamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata> loadResource = client.load(IoUtil.toUtf8Stream(modelYamlStr));
             if (!isApplyTask()) {
-                loadResource.delete();
+                if (roleType.equalsIgnoreCase("HELM_CHART")) {
+                    HasMetadata firstResource = loadResource.get().get(0);
+                    // 如果helmChart不存在，那么不会创建delete job，不能使用waitForJobCompleted
+                    if (firstResource == null) {
+                        loadResource.delete();
+                        loadResource.waitUntilCondition(Objects::isNull, 3, TimeUnit.MINUTES);
+                    } else {
+                        HelmChart helmChart = client.resources(HelmChart.class).load(IoUtil.toUtf8Stream(modelYamlStr)).item();
+                        Job helmChartJob = K8sUtil.getDeleteJobByHelmChart(client, helmChart);
+                        K8sUtil.waitForJobCompleted(() -> {
+                            loadResource.delete();
+                            loadResource.waitUntilCondition(Objects::isNull, 3, TimeUnit.MINUTES);
+                        }, taskParam, client, helmChartJob, 300);
+                    }
+                } else {
+                    loadResource.delete();
+                    loadResource.waitUntilCondition(Objects::isNull, 3, TimeUnit.MINUTES);
+                }
             } else {
                 try {
-                    List<HasMetadata> metadata = loadResource.forceConflicts().serverSideApply();
-                    resourceName = metadata.get(0).getMetadata().getName();
+                    long waitSeconds = 600;
                     switch (roleType.toUpperCase()) {
                         case "JOB":
-                            long waitSeconds = 300;
-                            K8sUtil.waitForJobCompleted(namespace, resourceName, client, log, waitSeconds);
+                            Job job = client.batch().v1().jobs().load(IoUtil.toUtf8Stream(modelYamlStr)).item();
+                            K8sUtil.waitForJobCompleted(() -> loadResource.forceConflicts().serverSideApply(), taskParam, client, job, waitSeconds);
+                            break;
+                        case "HELM_CHART":
+                            HelmChart helmChart = client.resources(HelmChart.class).load(IoUtil.toUtf8Stream(modelYamlStr)).item();
+                            Job helmChartInstallJob = K8sUtil.getInstallJobByHelmChart(client, helmChart);
+                            K8sUtil.waitForJobCompleted(() -> loadResource.forceConflicts().serverSideApply(), taskParam, client, helmChartInstallJob, waitSeconds);
+                            break;
+                        case "DEPLOYMENT":
+                            Deployment deployment = client.apps().deployments().load(IoUtil.toUtf8Stream(modelYamlStr)).item();
+                            K8sUtil.waitForDeploymentReady(() -> loadResource.forceConflicts().serverSideApply(), taskParam, client, deployment, waitSeconds);
                             break;
                         default:
-                            final Deployment deployment = client.apps().deployments().withName(resourceName).get();
-                            Resource<Deployment> resource = client.resource(deployment);
-                            int amount = 10;
-                            log.info("在k8s上启动deployment: {}  ,持续等待 {} 分钟", resourceName, amount);
-                            resource.waitUntilReady(amount, TimeUnit.MINUTES);
-
-                            // 打印deployment的输出日志
-                            RollableScalableResource<Deployment> scalableResource = client.apps().deployments().withName(resourceName);
-                            log.info(scalableResource.getLog());
+                            loadResource.forceConflicts().serverSideApply();
+                            loadResource.waitUntilReady(waitSeconds, TimeUnit.SECONDS);
                     }
                 } catch (Exception e) {
                     log.error(e.getMessage() + ":\n" + modelYamlStr);
-                    if (!"JOB".equalsIgnoreCase(roleType)) {
-                        // 打印deployment的输出日志
-                        if (StringUtils.isNotEmpty(resourceName)) {
-                            RollableScalableResource<Deployment> scalableResource = client.apps().deployments().withName(resourceName);
-                            log.error(scalableResource.getLog());
-                        }
-                    }
                     throw new RuntimeException(e);
                 }
             }
